requirements.txt
flask==2.0.3
flask-SQLAlchemy==2.5.1
ariadne==0.14.1
flask-Cors==3.0.10
SQLAlchemy-Utils==0.38.2
PyMySQL==1.0.2